pyproject.toml
[project]
name = "a2a-sample-agent-langgraph"
version = "0.1.0"
description = "LangGraph agent with A2A Protocol and helpfulness evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "xmltodict>=0.14.2",
    "chainlit>=2.6.9",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langgraph>=0.3.18",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "a2a-sdk>=0.3.0",
    "tavily-python>=0.3.0",
    "arxiv>=2.1.0",
    "tiktoken>=0.5.1",
    "qdrant-client>=1.7.0",
    "pymupdf>=1.23.0",
    "chainlit>=2.7.2",
    "langchain-tavily>=0.2.11",
    "flask>=3.1.2",
    "twilio>=9.8.2",
    "openmeteo-requests>=1.7.3",
    "pandas>=2.3.3",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
